refactor(runtime-core): 2.5c - extract PermissionRouter
Lift the permission orchestration responsibility out of HappyRuntimeStore
(runtime/happyClient.ts) into a dedicated PermissionRouter class under
runtime/orchestration/. Behavior preserved 1:1.

What moves:
- 6 in-memory Maps:
  permissions, codexPermissionIndex, codexPermissionResponders,
  providerPermissionIndex, providerPermissionWaiters,
  providerPermissionDecisions.
- 5 private methods:
  finalizeCodexRuntimePermissions, waitForPersistedPermissionDecision,
  awaitProviderPermissionDecision, watchExternalPermissionDecision,
  handleProviderPermissionRequest.
- persistPermissionEvent (private) is folded into the router as well.
- buildScopedPermissionKey helper (key format unchanged).
- HappyRuntimePermissionInput / RuntimeCoordinationStore types move to
  runtime/contracts/runtimeCoordinationStore.ts so orchestration modules
  can consume them without circular import via happyClient.

What stays on HappyRuntimeStore:
- listPermissions / createPermission / decidePermission as thin
  delegating wrappers, preserving the external surface used by
  store.ts / server.ts handlers.
- coordinationStore field (also passed into the router).

Wiring:
- Constructor builds PermissionRouter once with bound host callbacks:
  getSession, resolveApprovalPolicy, abortSessionRuns,
  appendAgentMessage, appendRunLifecycleEvent.
- Codex registerPermissionResponder lambda, enqueuePermission lambda,
  finalize call site, and claude/gemini onPermission hooks now call
  permissionRouter directly.
- happyClientTestHooks.finalizeCodexRuntimePermissions delegates to
  router via typed cast.

Behavior preservation guards:
- buildScopedPermissionKey format (chatId:baseKey) preserved verbatim.
- HappyRuntimePermissionInput.chatId still string | null | undefined.
- decidePermission ordering preserved: persisted update -> permissions
  cache update -> persistPermissionEvent -> codex/provider index
  cleanup -> decision cache -> provider waiter resolve -> codex
  responder dispatch -> abortSessionRuns on deny.
- happy event log payload keys unchanged.

Test surface:
- 1 test in happyClient.streamJson.test.ts updated to reach into
  store.permissionRouter.{permissions,codexPermissionIndex,
  codexPermissionResponders} instead of the now-removed top-level
  fields.

Validation:
- npx tsc --noEmit clean.
- npx vitest run --exclude e2e: 261/261 PASS (regression 0).

Net: 4 files, +565 -354. happyClient.ts alone -334 lines (1907 -> 1573).

Refs Phase 2.5 sprint 2.5c of
docs/03-platform/runtime-core-extraction-plan.md.
