+++ services/aris-backend/src/runtime/happyClient.ts
@@ -38,6 +38,11 @@ import type {
   ProviderRuntimeSession,
   ProviderTextEvent,
 } from './contracts/providerRuntime.js';
+import type {
+  HappyRuntimePermissionInput,
+  RuntimeCoordinationStore,
+} from './contracts/runtimeCoordinationStore.js';
+import { PermissionRouter, buildScopedPermissionKey } from './orchestration/permissionRouter.js';
 import type { ClaudeSessionLaunchMode } from './providers/claude/claudeSessionContract.js';
 import type { ClaudeActionEvent, ClaudeLaunchCommand, ClaudeResumeTarget, ClaudeTextEvent } from './providers/claude/types.js';
 import type {
@@ -169,28 +174,6 @@ type HappyRuntimeAppendInput = {
   title?: string;
   text: string;
   meta?: Record<string, unknown>;
-};
-
-type HappyRuntimePermissionInput = {
-  sessionId: string;
-  chatId?: string | null;
-  agent: PermissionRequest['agent'];
-  command: string;
-  reason: string;
-  risk: PermissionRisk;
-};
-
-type RuntimeCoordinationStore = {
-  listPermissions(state?: PermissionState): Promise<PermissionRequest[]>;
-  createPermission(input: HappyRuntimePermissionInput): Promise<PermissionRequest>;
-  decidePermission(permissionId: string, decision: PermissionDecision): Promise<PermissionRequest>;
-  getPermissionById(permissionId: string): Promise<PermissionRequest | null>;
-  hasRequestedAction(input: {
-    sessionId: string;
-    action: SessionAction;
-    chatId?: string;
-    createdAfter?: Date;
-  }): Promise<boolean>;
 };
 
 type SessionRealtimeEventRecord = {
@@ -1354,13 +1337,6 @@ function buildCodexPermissionKey(sessionId: string, request: CodexPermissionRequ
   return `${sessionId}:${request.approvalId || request.callId}`;
 }
 
-function buildScopedPermissionKey(baseKey: string, chatId?: string): string {
-  const normalizedChatId = typeof chatId === 'string' && chatId.trim().length > 0
-    ? chatId.trim()
-    : '__default__';
-  return `${normalizedChatId}:${baseKey}`;
-}
-
 function resolveClaudeLaunchMode(input: {
   sessionPath?: string;
   workspaceRoot: string;
@@ -1913,28 +1889,21 @@ export const happyClientTestHooks = {
     store: HappyRuntimeStore,
     permissionIds: Iterable<string>,
     options?: { preservePending?: boolean },
-  ) => (store as any).finalizeCodexRuntimePermissions(permissionIds, options),
+  ) => (store as unknown as { permissionRouter: PermissionRouter }).permissionRouter
+    .finalizeCodexPermissions(permissionIds, options),
 };
 
 export class HappyRuntimeStore {
-  private readonly permissions = new Map<string, PermissionRequest>();
   private readonly activeRuns = new Map<string, ActiveRun>();
   private readonly claudeSessionRegistry = new ClaudeSessionRegistry();
   private readonly geminiSessionRegistry = new GeminiSessionRegistry();
   private readonly claudeSessionScanners = new Map<string, ClaudeSessionLogTracker>();
   private readonly codexThreads = new Map<string, string>();
-  private readonly codexPermissionIndex = new Map<string, string>();
-  private readonly codexPermissionResponders = new Map<string, (decision: PermissionDecision) => Promise<void>>();
-  private readonly providerPermissionIndex = new Map<string, string>();
-  private readonly providerPermissionWaiters = new Map<string, {
-    resolve: (decision: PermissionDecision) => void;
-    reject: (error: Error) => void;
-  }>();
-  private readonly providerPermissionDecisions = new Map<string, PermissionDecision>();
   private readonly sessionRealtimeEvents = new Map<string, SessionRealtimeEventRecord[]>();
   private readonly sessionRealtimeCursor = new Map<string, number>();
   private readonly geminiPartialTextStates = new Map<string, GeminiPartialTextState>();
   private readonly coordinationStore: RuntimeCoordinationStore | null;
+  private readonly permissionRouter: PermissionRouter;
 
   private readonly serverUrl: string;
   private readonly serverToken: string;
@@ -1960,6 +1929,16 @@ export class HappyRuntimeStore {
     this.hostProjectsRoot = (opts.hostProjectsRoot || '').replace(/\/+$/, '');
     this.coordinationStore = opts.coordinationStore ?? null;
     this.happyEventLogger = new HappyEventLogger(HAPPY_EVENT_LOG_DIR, HAPPY_EVENT_LOG_MAX_BYTES);
+    this.permissionRouter = new PermissionRouter({
+      coordinationStore: this.coordinationStore,
+      getSession: (sessionId) => this.getSession(sessionId),
+      resolveApprovalPolicy: (session) => this.resolveSessionApprovalPolicy(session),
+      abortSessionRuns: (sessionId, chatId) => this.abortSessionRuns(sessionId, chatId),
+      appendAgentMessage: (sessionId, text, meta, options) =>
+        this.appendAgentMessage(sessionId, text, meta, options),
+      appendRunLifecycleEvent: (sessionId, state, meta) =>
+        this.appendRunLifecycleEvent(sessionId, state, meta),
+    });
   }
 
   beginShutdownDrain(): void {
@@ -1978,38 +1957,6 @@ export class HappyRuntimeStore {
 
   private getActiveRunCount(): number {
     return this.activeRuns.size + this.claudeSessionRegistry.activeRunCount();
-  }
-
-  private async finalizeCodexRuntimePermissions(
-    permissionIds: Iterable<string>,
-    options: { preservePending?: boolean } = {},
-  ): Promise<void> {
-    for (const permissionId of permissionIds) {
-      this.codexPermissionResponders.delete(permissionId);
-
-      for (const [key, mappedPermissionId] of this.codexPermissionIndex.entries()) {
-        if (mappedPermissionId === permissionId) {
-          this.codexPermissionIndex.delete(key);
-        }
-      }
-
-      const existing = this.permissions.get(permissionId);
-      if (!existing || existing.state !== 'pending' || options.preservePending) {
-        continue;
-      }
-
-      if (this.coordinationStore) {
-        try {
-          const denied = await this.coordinationStore.decidePermission(permissionId, 'deny');
-          this.permissions.set(permissionId, denied);
-          continue;
-        } catch {
-          // Fall through to the in-memory copy when the persisted store is unavailable.
-        }
-      }
-
-      this.permissions.set(permissionId, { ...existing, state: 'denied', decision: 'deny' });
-    }
   }
 
   private resolveSessionApprovalPolicy(session: RuntimeSession): ApprovalPolicy {
@@ -2162,101 +2109,6 @@ export class HappyRuntimeStore {
     }
   }
 
-  private async waitForPersistedPermissionDecision(
-    permissionId: string,
-    signal?: AbortSignal,
-  ): Promise<PermissionDecision> {
-    if (!this.coordinationStore) {
-      throw new Error('PERSISTED_PERMISSION_COORDINATION_UNAVAILABLE');
-    }
-
-    while (true) {
-      if (signal?.aborted) {
-        throw new Error('The operation was aborted');
-      }
-
-      const persisted = await this.coordinationStore.getPermissionById(permissionId);
-      if (!persisted) {
-        throw new Error('PERMISSION_NOT_FOUND');
-      }
-      this.permissions.set(permissionId, persisted);
-      if (persisted.state === 'denied') {
-        return 'deny';
-      }
-      if (persisted.state === 'approved') {
-        return persisted.decision === 'allow_session' ? 'allow_session' : 'allow_once';
-      }
-
-      await new Promise((resolve) => setTimeout(resolve, 400));
-    }
-  }
-
-  private async awaitProviderPermissionDecision(
-    permissionId: string,
-    signal?: AbortSignal,
-  ): Promise<PermissionDecision> {
-    if (this.coordinationStore) {
-      return this.waitForPersistedPermissionDecision(permissionId, signal);
-    }
-
-    const knownDecision = this.providerPermissionDecisions.get(permissionId);
-    if (knownDecision) {
-      return knownDecision;
-    }
-
-    const existing = this.permissions.get(permissionId);
-    if (existing && existing.state !== 'pending') {
-      return existing.state === 'denied' ? 'deny' : 'allow_once';
-    }
-
-    return new Promise<PermissionDecision>((resolve, reject) => {
-      const handleAbort = () => {
-        this.providerPermissionWaiters.delete(permissionId);
-        reject(new Error('The operation was aborted'));
-      };
-
-      if (signal?.aborted) {
-        handleAbort();
-        return;
-      }
-
-      const cleanup = () => {
-        signal?.removeEventListener('abort', handleAbort);
-      };
-
-      this.providerPermissionWaiters.set(permissionId, {
-        resolve: (decision) => {
-          cleanup();
-          resolve(decision);
-        },
-        reject: (error) => {
-          cleanup();
-          reject(error);
-        },
-      });
-      signal?.addEventListener('abort', handleAbort, { once: true });
-    });
-  }
-
-  private watchExternalPermissionDecision(input: {
-    permissionId: string;
-    responder: (decision: PermissionDecision) => Promise<void>;
-    signal?: AbortSignal;
-  }): void {
-    if (!this.coordinationStore) {
-      return;
-    }
-
-    void this.waitForPersistedPermissionDecision(input.permissionId, input.signal)
-      .then((decision) => input.responder(decision))
-      .catch((error) => {
-        const message = error instanceof Error ? error.message : String(error);
-        if (!message.includes('aborted')) {
-          console.error(`failed to resolve persisted permission decision: ${message}`);
-        }
-      });
-  }
-
   private monitorExternalAbortSignal(input: {
     sessionId: string;
     chatId?: string;
@@ -2300,48 +2152,6 @@ export class HappyRuntimeStore {
     return () => {
       disposed = true;
     };
-  }
-
-  private async handleProviderPermissionRequest(input: {
-    session: RuntimeSession;
-    chatId?: string;
-    agent: PermissionRequest['agent'];
-    request: ProviderPermissionRequest;
-    signal?: AbortSignal;
-  }): Promise<PermissionDecision> {
-    const permissionKey = buildScopedPermissionKey(
-      `${input.session.id}:provider:${input.request.approvalId || input.request.callId}`,
-      input.chatId,
-    );
-    const knownPermissionId = this.providerPermissionIndex.get(permissionKey);
-    if (knownPermissionId) {
-      const knownPermission = this.permissions.get(knownPermissionId);
-      if (knownPermission?.state === 'pending') {
-        return this.awaitProviderPermissionDecision(knownPermissionId, input.signal);
-      }
-      if (knownPermission) {
-        return knownPermission.state === 'denied' ? 'deny' : 'allow_once';
-      }
-      this.providerPermissionIndex.delete(permissionKey);
-    }
-
-    const created = await this.createPermission({
-      sessionId: input.session.id,
-      ...(input.chatId ? { chatId: input.chatId } : {}),
-      agent: input.agent,
-      command: input.request.command,
-      reason: input.request.reason,
-      risk: input.request.risk,
-    });
-    this.providerPermissionIndex.set(permissionKey, created.id);
-
-    const approvalPolicy = this.resolveSessionApprovalPolicy(input.session);
-    if (approvalPolicy === 'yolo') {
-      await this.decidePermission(created.id, 'allow_session');
-      return 'allow_session';
-    }
-
-    return this.awaitProviderPermissionDecision(created.id, input.signal);
   }
 
   private getClaudeSessionScanner(workingDirectory: string): ClaudeSessionLogTracker {
@@ -3514,18 +3324,18 @@ export class HappyRuntimeStore {
       risk: PermissionRisk,
       responder: (decision: PermissionDecision) => Promise<void>,
     ) => {
-      const knownPermissionId = this.codexPermissionIndex.get(key);
+      const knownPermissionId = this.permissionRouter.lookupCodexBinding(key);
       if (knownPermissionId) {
-        const knownPermission = this.permissions.get(knownPermissionId);
+        const knownPermission = this.permissionRouter.getCachedPermission(knownPermissionId);
         if (knownPermission?.state === 'pending') {
           if (this.coordinationStore) {
-            this.watchExternalPermissionDecision({
+            this.permissionRouter.watchExternal({
               permissionId: knownPermissionId,
               responder,
               signal,
             });
           } else {
-            this.codexPermissionResponders.set(knownPermissionId, responder);
+            this.permissionRouter.registerCodexResponder(knownPermissionId, responder);
           }
           runtimePermissionIds.add(knownPermissionId);
           if (autoApproveAll) {
@@ -3533,7 +3343,7 @@ export class HappyRuntimeStore {
           }
           return;
         }
-        this.codexPermissionIndex.delete(key);
+        this.permissionRouter.clearCodexBinding(key);
       }
 
       const created = await this.createPermission({
@@ -3545,15 +3355,15 @@ export class HappyRuntimeStore {
         risk,
       });
 
-      this.codexPermissionIndex.set(key, created.id);
+      this.permissionRouter.registerCodexBinding(key, created.id);
       if (this.coordinationStore) {
-        this.watchExternalPermissionDecision({
+        this.permissionRouter.watchExternal({
           permissionId: created.id,
           responder,
           signal,
         });
       } else {
-        this.codexPermissionResponders.set(created.id, responder);
+        this.permissionRouter.registerCodexResponder(created.id, responder);
       }
       runtimePermissionIds.add(created.id);
       if (autoApproveAll) {
@@ -4133,7 +3943,7 @@ export class HappyRuntimeStore {
       await permissionChain.catch(() => undefined);
       await appendChain.catch(() => undefined);
 
-      await this.finalizeCodexRuntimePermissions(runtimePermissionIds, {
+      await this.permissionRouter.finalizeCodexPermissions(runtimePermissionIds, {
         preservePending: this.draining && !signal?.aborted,
       });
 
@@ -4284,16 +4094,16 @@ export class HappyRuntimeStore {
       permissionChain = permissionChain
         .then(async () => {
           const key = buildScopedPermissionKey(buildCodexPermissionKey(session.id, request), chatId);
-          const knownPermissionId = this.codexPermissionIndex.get(key);
+          const knownPermissionId = this.permissionRouter.lookupCodexBinding(key);
           if (knownPermissionId) {
-            const knownPermission = this.permissions.get(knownPermissionId);
+            const knownPermission = this.permissionRouter.getCachedPermission(knownPermissionId);
             if (knownPermission?.state === 'pending') {
               if (autoApproveAll) {
                 await this.decidePermission(knownPermissionId, 'allow_session');
               }
               return;
             }
-            this.codexPermissionIndex.delete(key);
+            this.permissionRouter.clearCodexBinding(key);
           }
 
           const created = await this.createPermission({
@@ -4305,7 +4115,7 @@ export class HappyRuntimeStore {
             risk: request.risk,
           });
 
-          this.codexPermissionIndex.set(key, created.id);
+          this.permissionRouter.registerCodexBinding(key, created.id);
           if (autoApproveAll) {
             await this.decidePermission(created.id, 'allow_session');
           }
@@ -4942,7 +4752,7 @@ export class HappyRuntimeStore {
                 envelopes: event.envelopes,
               });
             },
-            onPermission: async (request) => this.handleProviderPermissionRequest({
+            onPermission: async (request) => this.permissionRouter.handleProviderPermissionRequest({
               session,
               chatId: scopedChatId,
               agent: 'claude',
@@ -5069,7 +4879,7 @@ export class HappyRuntimeStore {
               });
               await geminiMessageQueue?.flush();
             },
-            onPermission: async (request) => this.handleProviderPermissionRequest({
+            onPermission: async (request) => this.permissionRouter.handleProviderPermissionRequest({
               session,
               chatId: scopedChatId,
               agent: 'gemini',
@@ -5789,133 +5599,14 @@ export class HappyRuntimeStore {
   }
 
   async listPermissions(state?: PermissionState): Promise<PermissionRequest[]> {
-    if (this.coordinationStore) {
-      const persisted = await this.coordinationStore.listPermissions(state);
-      for (const permission of persisted) {
-        this.permissions.set(permission.id, permission);
-      }
-      return persisted;
-    }
-    const list = [...this.permissions.values()].sort((a, b) => b.requestedAt.localeCompare(a.requestedAt));
-    return state ? list.filter((permission) => permission.state === state) : list;
+    return this.permissionRouter.listPermissions(state);
   }
 
   async createPermission(input: HappyRuntimePermissionInput): Promise<PermissionRequest> {
-    const session = await this.getSession(input.sessionId);
-    if (!session) {
-      throw new Error('SESSION_NOT_FOUND');
-    }
-
-    const permission = this.coordinationStore
-      ? await this.coordinationStore.createPermission(input)
-      : {
-          id: randomUUID(),
-          sessionId: input.sessionId,
-          ...(typeof input.chatId === 'string' && input.chatId.trim().length > 0
-            ? { chatId: input.chatId.trim() }
-            : {}),
-          agent: input.agent,
-          command: input.command,
-          reason: input.reason,
-          risk: input.risk,
-          requestedAt: new Date().toISOString(),
-          state: 'pending' as const,
-        };
-
-    this.permissions.set(permission.id, permission);
-    await this.persistPermissionEvent(permission, 'permission_request');
-    await this.appendRunLifecycleEvent(input.sessionId, 'waiting_for_approval', {
-      ...(permission.chatId ? { chatId: permission.chatId } : {}),
-      requestedPath: session.metadata.path,
-      agent: input.agent,
-      command: input.command,
-      reason: input.reason,
-    });
-    return permission;
+    return this.permissionRouter.createPermission(input);
   }
 
   async decidePermission(permissionId: string, decision: PermissionDecision): Promise<PermissionRequest> {
-    const knownPermission = this.permissions.get(permissionId);
-    const permission = knownPermission
-      ?? await this.coordinationStore?.getPermissionById(permissionId)
-      ?? null;
-    if (!permission) {
-      throw new Error('PERMISSION_NOT_FOUND');
-    }
-
-    const updated = this.coordinationStore
-      ? await this.coordinationStore.decidePermission(permissionId, decision)
-      : {
-          ...permission,
-          state: (decision === 'deny' ? 'denied' : 'approved') as PermissionState,
-          decision,
-        };
-    this.permissions.set(permissionId, updated);
-    await this.persistPermissionEvent(updated, 'permission_decision', decision);
-
-    for (const [key, mappedPermissionId] of this.codexPermissionIndex.entries()) {
-      if (mappedPermissionId === permissionId) {
-        this.codexPermissionIndex.delete(key);
-      }
-    }
-    for (const [key, mappedPermissionId] of this.providerPermissionIndex.entries()) {
-      if (mappedPermissionId === permissionId) {
-        this.providerPermissionIndex.delete(key);
-      }
-    }
-
-    this.providerPermissionDecisions.set(permissionId, decision);
-    const providerWaiter = this.providerPermissionWaiters.get(permissionId);
-    this.providerPermissionWaiters.delete(permissionId);
-    providerWaiter?.resolve(decision);
-
-    const responder = this.codexPermissionResponders.get(permissionId);
-    this.codexPermissionResponders.delete(permissionId);
-
-    if (responder) {
-      try {
-        await responder(decision);
-      } catch (error) {
-        const message = error instanceof Error ? error.message : String(error);
-        console.error(`failed to send codex permission decision: ${message}`);
-        if (decision === 'deny') {
-          this.abortSessionRuns(permission.sessionId, permission.chatId ?? undefined);
-        }
-      }
-    } else if (decision === 'deny') {
-      this.abortSessionRuns(permission.sessionId, permission.chatId ?? undefined);
-    }
-
-    return updated;
-  }
-
-  private async persistPermissionEvent(
-    permission: PermissionRequest,
-    streamEvent: 'permission_request' | 'permission_decision',
-    decision?: PermissionDecision,
-  ): Promise<void> {
-    try {
-      await this.appendAgentMessage(
-        permission.sessionId,
-        permission.command,
-        {
-          sessionId: permission.sessionId,
-          ...(permission.chatId ? { chatId: permission.chatId } : {}),
-          agent: permission.agent,
-          streamEvent,
-          permissionId: permission.id,
-          permissionState: permission.state,
-          command: permission.command,
-          reason: permission.reason,
-          risk: permission.risk,
-          requestedAt: permission.requestedAt,
-          ...(decision ? { permissionDecision: decision } : {}),
-        },
-        { type: 'message', title: 'Permission Request' },
-      );
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      console.error(`failed to persist permission event (${streamEvent}): ${message}`);
-    }
+    return this.permissionRouter.decidePermission(permissionId, decision);
   }
 }
